requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile setup.py
#
absl-py==0.9.0            # via nanoemoji (setup.py)
appdirs==1.4.4            # via fs
attrs==19.3.0             # via ufolib2
booleanoperations==0.9.0  # via ufo2ft
cffsubr==0.2.7            # via ufo2ft
compreffor==0.5.0         # via ufo2ft
cu2qu==1.6.7              # via ufo2ft
fonttools[ufo]==4.13.0    # via booleanoperations, cffsubr, compreffor, cu2qu, nanoemoji (setup.py), ufo2ft, ufolib2
fs==2.4.11                # via fonttools
lxml==4.5.1               # via nanoemoji (setup.py), picosvg
ninja==1.10.0.post1       # via nanoemoji (setup.py)
picosvg==0.5.0            # via nanoemoji (setup.py)
pyclipper==1.2.0    # via booleanoperations
pytz==2020.1              # via fs
regex==2020.6.8           # via nanoemoji (setup.py)
six==1.15.0               # via absl-py, fs
skia-pathops==0.4.1       # via picosvg
ufo2ft[cffsubr]==2.15.0   # via nanoemoji (setup.py)
ufolib2==0.6.2            # via nanoemoji (setup.py)

# The following packages are considered to be unsafe in a requirements file:
# setuptools
